+++ geotrek/feedback/tests/test_forms.py
@@ -3,7 +3,7 @@ from hashlib import md5
 from unittest import mock
 
 from django.core import mail
-from django.forms.widgets import EmailInput, HiddenInput, Select
+from django.forms.widgets import CheckboxInput, EmailInput, HiddenInput, Select
 from django.test.utils import override_settings
 from django.urls.base import reverse
 from mapentity.tests.factories import SuperUserFactory
@@ -27,7 +27,8 @@ class TestSuricateForms(SuricateWorkflowTests):
     def setUp(cls):
         super().setUp()
         cls.filed_report = ReportFactory(status=cls.filed_status, uid=uuid.uuid4())
-        cls.waiting_report = ReportFactory(status=cls.waiting_status, uid=uuid.uuid4())
+        cls.waiting_report = ReportFactory(status=cls.waiting_status, uses_timers=True, uid=uuid.uuid4())
+        cls.intervention_solved_report = ReportFactory(status=cls.intervention_solved_status, uid=uuid.uuid4())
         cls.client.login(username="Admiin", password="drowssap")
 
     @test_for_report_and_basic_modes
@@ -46,8 +47,10 @@ class TestSuricateForms(SuricateWorkflowTests):
         self.assertIsInstance(form.fields["status"].widget, Select)
         self.assertIsInstance(form.fields["problem_magnitude"].widget, Select)
         self.assertIsInstance(form.fields["related_trek"].widget, Select)
-        self.assertNotIn('message', keys)
+        self.assertNotIn('message_sentinel', keys)
+        self.assertNotIn('message_supervisor', keys)
         self.assertIsInstance(form.fields["assigned_user"].widget, HiddenInput)
+        self.assertIsInstance(form.fields["uses_timers"].widget, HiddenInput)
         self.assertFalse(form.errors)
 
     @test_for_report_and_basic_modes
@@ -62,8 +65,10 @@ class TestSuricateForms(SuricateWorkflowTests):
         self.assertIsInstance(form.fields["status"].widget, Select)
         self.assertIsInstance(form.fields["problem_magnitude"].widget, Select)
         self.assertIsInstance(form.fields["related_trek"].widget, Select)
-        self.assertNotIn('message', keys)
+        self.assertNotIn('message_sentinel', keys)
+        self.assertNotIn('message_supervisor', keys)
         self.assertIsInstance(form.fields["assigned_user"].widget, HiddenInput)
+        self.assertIsInstance(form.fields["uses_timers"].widget, HiddenInput)
         self.assertFalse(form.errors)  # assert form is valid
 
     @test_for_management_mode
@@ -83,8 +88,10 @@ class TestSuricateForms(SuricateWorkflowTests):
         self.assertIsInstance(form.fields["status"].widget, HiddenInput)
         self.assertIsInstance(form.fields["problem_magnitude"].widget, Select)
         self.assertIsInstance(form.fields["related_trek"].widget, Select)
-        self.assertNotIn('message', keys)
+        self.assertNotIn('message_sentinel', keys)
+        self.assertNotIn('message_supervisor', keys)
         self.assertIsInstance(form.fields["assigned_user"].widget, HiddenInput)
+        self.assertIsInstance(form.fields["uses_timers"].widget, HiddenInput)
 
     @test_for_management_mode
     def test_update_form_specifics_2(self):
@@ -98,8 +105,10 @@ class TestSuricateForms(SuricateWorkflowTests):
         self.assertIsInstance(form.fields["status"].widget, Select)
         self.assertIsInstance(form.fields["problem_magnitude"].widget, HiddenInput)
         self.assertIsInstance(form.fields["related_trek"].widget, Select)
-        self.assertIn('message', keys)
+        self.assertIn('message_sentinel', keys)
+        self.assertIn('message_supervisor', keys)
         self.assertIsInstance(form.fields["assigned_user"].widget, Select)
+        self.assertIsInstance(form.fields["uses_timers"].widget, CheckboxInput)
         # Todo ajouter les contraintes de contenu de status selon old_status / pas de contrainte si autres modes
 
     @test_for_management_mode
@@ -114,28 +123,34 @@ class TestSuricateForms(SuricateWorkflowTests):
             'assigned_user': str(self.user.pk),
             'email': 'test@test.fr',
             'geom': 'POINT(5.1 6.6)',
+            'message_sentinel': "Your message",
+            "uses_timers": True
         }
         form = ReportForm(instance=self.filed_report, data=data)
         form.save()
         # Assert report status changes
         self.assertEquals(self.filed_report.status.suricate_id, "waiting")
         # Asser timer is created
-        self.assertEquals(TimerEvent.objects.filter(report=self.filed_report).count(), 1)
+        self.assertEquals(TimerEvent.objects.filter(report=self.filed_report, step=self.waiting_status).count(), 1)
         # Assert data forwarded to Suricate
         check = md5(
             (SuricateMessenger().gestion_manager.PRIVATE_KEY_CLIENT_SERVER + SuricateMessenger().gestion_manager.ID_ORIGIN + str(self.filed_report.uid)).encode()
         ).hexdigest()
         call1 = mock.call(
             'http://suricate.example.com/wsSendMessageSentinelle',
-            {'id_origin': 'geotrek', 'uid_alerte': self.filed_report.uid, 'message': '', 'check': check},
+            {'id_origin': 'geotrek', 'uid_alerte': self.filed_report.uid, 'message': 'Your message', 'check': check},
             auth=('', '')
         )
         call2 = mock.call(
             'http://suricate.example.com/wsUpdateStatus',
-            {'id_origin': 'geotrek', 'uid_alerte': self.filed_report.uid, 'statut': 'waiting', 'txt_changestatut': '', 'check': check},
+            {'id_origin': 'geotrek', 'uid_alerte': self.filed_report.uid, 'statut': 'waiting', 'txt_changestatut': 'Your message', 'check': check},
             auth=('', '')
         )
         mocked_post.assert_has_calls([call1, call2], any_order=True)
+        mocked_get.assert_called_once_with(
+            f"http://suricate.example.com/wsLockAlert?id_origin=geotrek&uid_alerte={self.filed_report.uid}&check={check}",
+            auth=('', '')
+        )
         # Assert user is notified
         self.assertEqual(len(mail.outbox), mails_before + 1)
         self.assertEqual(mail.outbox[-1].subject, "Geotrek - Nouveau Signalement à traiter")
@@ -151,7 +166,7 @@ class TestSuricateForms(SuricateWorkflowTests):
             'name': "test_interv",
             'date': "2025-12-12",
             'status': status.pk,
-            'structure': user.profile.structure.pk
+            'structure': user.profile.structure.pk,
         }
         form = InterventionForm(user=user, target_type=self.waiting_report.get_content_type_id(), target_id=self.waiting_report.pk, data=data)
         form.is_valid()
@@ -206,3 +221,40 @@ class TestSuricateForms(SuricateWorkflowTests):
         response = self.client.get(reverse('feedback:report_detail', kwargs={'pk': self.waiting_report.pk}), follow=True)
         self.assertEquals(response.status_code, 200)
         self.assertIn("Add a new intervention", response.content.decode("utf-8"))
+
+    @test_for_management_mode
+    @mock.patch("geotrek.feedback.helpers.requests.get")
+    @mock.patch("geotrek.feedback.helpers.requests.post")
+    def test_workflow_resolve_step(self, mocked_post, mocked_get):
+        self.build_get_request_patch(mocked_get)
+        self.build_post_request_patch(mocked_post)
+        # When assigning a user to a report
+        data = {
+            'email': 'test@test.fr',
+            'geom': 'POINT(5.1 6.6)',
+            'status': self.resolved_status.pk,
+            'message_sentinel': "Your message"
+        }
+        form = ReportForm(instance=self.intervention_solved_report, data=data)
+        form.save()
+        # Assert report status changes
+        self.assertEquals(self.intervention_solved_report.status.suricate_id, "resolved")
+        # Assert data forwarded to Suricate
+        check = md5(
+            (SuricateMessenger().gestion_manager.PRIVATE_KEY_CLIENT_SERVER + SuricateMessenger().gestion_manager.ID_ORIGIN + str(self.intervention_solved_report.uid)).encode()
+        ).hexdigest()
+        call1 = mock.call(
+            'http://suricate.example.com/wsSendMessageSentinelle',
+            {'id_origin': 'geotrek', 'uid_alerte': self.intervention_solved_report.uid, 'message': 'Your message', 'check': check},
+            auth=('', '')
+        )
+        call2 = mock.call(
+            'http://suricate.example.com/wsUpdateStatus',
+            {'id_origin': 'geotrek', 'uid_alerte': self.intervention_solved_report.uid, 'statut': 'resolved', 'txt_changestatut': 'Your message', 'check': check},
+            auth=('', '')
+        )
+        mocked_post.assert_has_calls([call1, call2], any_order=True)
+        mocked_get.assert_called_once_with(
+            f"http://suricate.example.com/wsUnlockAlert?id_origin=geotrek&uid_alerte={self.intervention_solved_report.uid}&check={check}",
+            auth=('', '')
+        )